% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermoji_gadget.R
\name{emoji2html}
\alias{emoji2html}
\title{Convert Emoji to HTML}
\usage{
emoji2html(x, copy = clipr::clipr_available())
}
\arguments{
\item{x}{A search term passed to \code{\link[emo:ji]{emo::ji()}} or the result of \code{\link[emo:ji]{emo::ji()}}.}

\item{copy}{Should the result be copied to the clipboard?}
}
\description{
Uses \code{\link[emo:ji]{emo::ji()}} to look up emoji, but return the escaped HTML.
}
