% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_picker.R
\name{emoji_picker}
\alias{emoji_picker}
\title{Emoji Picker}
\usage{
emoji_picker(gadget = TRUE)
}
\arguments{
\item{gadget}{If \code{TRUE} the app is run as a gadget, otherwise it's run as a
standalone app.}
}
\value{
A list with the \code{emoji}, the emoji \code{name} and the \code{html} entity.
}
\description{
An emoji picker gadget that lets you search and find emoji and insert either
the unicode, HTML, or \code{emo::ji()} versions of the emoji. Built using the
\href{https://emoji-button.js.org/}{Emoji Button} JavaScript library.
}
