% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermoji_gadget.R
\name{ermoji}
\alias{ermoji}
\alias{ermoji_gadget}
\alias{ermoji_shiny}
\title{The ermoji emoji gadget}
\usage{
ermoji_gadget(clipout = clipr::clipr_available(), ...)

ermoji_shiny(clipout = clipr::clipr_available(), ...)
}
\arguments{
\item{clipout}{Should the gadget attempt to write to the clipboard?}

\item{...}{Ignored at this time}
}
\value{
nothing
}
\description{
Opens a miniUI based Shiny gadget in the RStudio Viewer pane with a
searchable table of emoji. Select a row and click the copy desired button.
}
