% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ermoji_gadget.R
\name{ermoji_gadget}
\alias{ermoji_gadget}
\title{The ermoji emoji gadget}
\usage{
ermoji_gadget()
}
\value{
nothing
}
\description{
Opens a miniUI based Shiny gadget in the RStudio Viewer pane with a
searchable table of emoji. Select a row and click the copy button.
}
